diff - moved transform methods to DiffTransformer
