+++ Wikipedia/Code/DiffListChangeItemViewModel.swift
@@ -29,19 +29,19 @@ final class DiffListChangeItemViewModel {
     
 
     init(item: TransformDiffItem, traitCollection: UITraitCollection, theme: Theme, type: DiffListChangeType, diffItemType: DiffItemType, nextMiddleItem: TransformDiffItem?, semanticContentAttribute: UISemanticContentAttribute) {
-        self.text = item.diffItem.text
+        self.text = item.text
 
         self.traitCollection = traitCollection
         self.theme = theme
         self.type = type
         self.diffItemType = diffItemType
 
-        self.moveInfo = item.transformMoveInfo
+        self.moveInfo = item.moveInfo
         self.semanticContentAttribute = semanticContentAttribute
         
         //account for utf8 offsets in highlighted ranges
         var convertedHighlightedRanges: [DiffHighlightRange] = []
-        if let diffHighlightedRanges = item.diffItem.highlightRanges {
+        if let diffHighlightedRanges = item.highlightRanges {
             for diffHighlightedRange in diffHighlightedRanges {
                 let start = diffHighlightedRange.start
                 let length = diffHighlightedRange.length
@@ -63,7 +63,7 @@ final class DiffListChangeItemViewModel {
         
         self.textPadding = DiffListChangeItemViewModel.calculateTextPadding(type: type, diffItemType: diffItemType, nextMiddleItem: nextMiddleItem)
 
-        self.textAttributedString = DiffListChangeItemViewModel.calculateAttributedString(with: text, highlightedRanges: highlightedRanges, traitCollection: traitCollection, theme: theme, type: type, diffItemType: diffItemType, moveInfo: item.transformMoveInfo, semanticContentAttribute: semanticContentAttribute)
+        self.textAttributedString = DiffListChangeItemViewModel.calculateAttributedString(with: text, highlightedRanges: highlightedRanges, traitCollection: traitCollection, theme: theme, type: type, diffItemType: diffItemType, moveInfo: item.moveInfo, semanticContentAttribute: semanticContentAttribute)
     }
     
     private static func calculateBackgroundColor(diffItemType: DiffItemType, theme: Theme) -> UIColor {
@@ -76,14 +76,14 @@ final class DiffListChangeItemViewModel {
         var bottom: CGFloat = 0
         if diffItemType == .moveSource || diffItemType == .moveDestination {
             top = 10
-            if let middleItem = nextMiddleItem?.diffItem,
+            if let middleItem = nextMiddleItem,
             middleItem.type == .moveSource || middleItem.type == .moveDestination {
                 bottom = 0
             } else {
                 bottom = 15
             }
         } else {
-            if let middleItem = nextMiddleItem?.diffItem,
+            if let middleItem = nextMiddleItem,
             middleItem.type == .moveSource || middleItem.type == .moveDestination {
                 bottom = 10
             }
@@ -113,7 +113,7 @@ final class DiffListChangeItemViewModel {
             
             let moveIndexString = "  \(moveIndex + 1)  "
             let imageAttachment = NSTextAttachment()
-            let imageName = moveInfo.moveInfo.linkDirection == .down ? "moveArrowDown" : "moveArrowUp"
+            let imageName = moveInfo.linkDirection == .down ? "moveArrowDown" : "moveArrowUp"
             imageAttachment.image = UIImage(named: imageName)
             let imageString = NSAttributedString(attachment: imageAttachment)
             
@@ -133,7 +133,7 @@ final class DiffListChangeItemViewModel {
                     
                     let paragraphMovedFormat = WMFLocalizedString("diff-paragraph-moved-format", value: "Paragraph moved %1$@ %2$@", comment: "Label in moved paragraph source location on diff view for indicating how far and what direction a pargraph has moved. %1$@ is replaced by the move direction (e.g. 'up' or 'down') and %2$@ is replaced by the move type and move distance (e.g. '2 lines', '4 sections')")
                     let upOrDownString: String
-                    switch moveInfo.moveInfo.linkDirection {
+                    switch moveInfo.linkDirection {
                     case .down: upOrDownString = WMFLocalizedString("diff-paragraph-moved-direction-down", value: "down", comment: "Label in moved pararaph source location on diff view for indicating that a paragraph was moved down in the document.")
                     case .up: upOrDownString = WMFLocalizedString("diff-paragraph-moved-direction-up", value: "up", comment: "Label in moved pararaph source location on diff view for indicating that a paragraph was moved up in the document.")
                     }
